+++ mopub-sdk/mopub-sdk-base/src/main/java/com/mopub/mobileads/AdRequestStatusMapping.java
@@ -5,7 +5,7 @@ import android.support.annotation.Nullable;
 import android.text.TextUtils;
 
 import com.mopub.common.Preconditions;
-import com.mopub.network.AdRequest;
+import com.mopub.common.VisibleForTesting;
 
 import java.util.Map;
 import java.util.TreeMap;
@@ -96,6 +96,12 @@ public class AdRequestStatusMapping {
         if (mAdUnitToAdRequestStatus.containsKey(adUnitId)) {
             mAdUnitToAdRequestStatus.get(adUnitId).setClickUrl(null);
         }
+    }
+
+    @Deprecated
+    @VisibleForTesting
+    void clearMapping() {
+        mAdUnitToAdRequestStatus.clear();
     }
 
     private static enum LoadingStatus { LOADING, LOADED, PLAYED }